fix(analytics-core): add support for experiment plugin
